requirements.txt
aiogram==3.16.0
pymongo==4.8.0
python-dotenv==1.0.0
apscheduler==3.11.0
pytz
